refactor(panel): split handlers.rs (1342 lines) into submodules
U-16 (<=800 lines/file) was violated. Move the ~40 axum handlers
into focused submodules grouped by route prefix and concern.
panel/mod.rs untouched: handlers/mod.rs re-exports all handlers
with pub(super) use, so the router wiring is unchanged.

- handlers/mod.rs          85
- shared.rs               106 (envelope helpers, OperationSummary)
- skill_read.rs           211 (SkillReadRow + build_skill_read_model)
- v1_routes.rs            234 (all v1_* endpoints)
- legacy.rs                84 (health, info, skills, etc.)
- registry_read.rs        200 (read-only registry endpoints)
- registry_mutations.rs   206 (target/binding/project/capture/orphan)
- skill_ops.rs            125 (skill add/save/snapshot/release/rollback)
- ops.rs                  128 (retry/purge/history_repair/pending)
- sync.rs                  72
- remote.rs                73
